Add new eligibility_criteria & create constants module
